# backend/docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  octotify_internal_handler_dto.ChannelDTO:
    description: 推送渠道的详细信息
    properties:
      config:
        description: 渠道配置，JSON 格式
        type: object
      created_at:
        description: 创建时间（Unix 毫秒时间戳）
        example: 1714636800000
        type: integer
      id:
        description: 渠道 ID
        example: 1
        type: integer
      last_used_at:
        description: 最后使用时间（Unix 毫秒时间戳，0 表示未使用）
        example: 1714636800000
        type: integer
      name:
        description: 渠道名称
        example: 钉钉-运维群
        type: string
      status:
        description: 状态：1-启用 2-停用 -1-已删除
        example: 1
        type: integer
      type:
        description: 渠道类型：wechat, telegram, dingtalk, email, webhook
        example: dingtalk
        type: string
      updated_at:
        description: 更新时间（Unix 毫秒时间戳）
        example: 1714636800000
        type: integer
      user_id:
        description: 所属用户 ID
        example: 1
        type: integer
    type: object
...
  octotify_internal_handler_dto.CreateSourceReq:
    description: 创建新的消息来源，系统自动生成 Source Token
    properties:
      channel_ids:
        description: 关联的渠道ID列表
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        type: array
      description:
        description: 来源描述，最多 512 个字符
        example: GitHub Actions 构建通知
        maxLength: 512
        type: string
      name:
        description: 来源名称，1-128 个字符
        example: CI Pipeline
        maxLength: 128
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  octotify_internal_handler_dto.SourceDTO:
    description: 消息来源的基本信息，不包含 Token
    properties:
      created_at:
        description: 创建时间（Unix 毫秒时间戳）
        example: 1714636800000
        type: integer
      description:
        description: 来源描述
        example: GitHub Actions 构建通知
        type: string
      id:
        description: 来源 ID
        example: 1
        type: integer
      name:
        description: 来源名称
        example: CI Pipeline
        type: string
      status:
        description: 状态：1-启用 2-停用 -1-已删除
        example: 1
        type: integer
      token:
        description: 推送 Token（仅创建时返回）
        example: src0196a3b2c4d50000a1b2c3d4e5f67890
        type: string
      user_id:
        description: 所属用户 ID
        example: 1
        type: integer
    type: object
  octotify_internal_handler_dto.SourceDetailDTO:
    description: 消息来源的详细信息，包含 Token 和使用时间
    properties:
      created_at:
        description: 创建时间（Unix 毫秒时间戳）
        example: 1714636800000
        type: integer
      description:
        description: 来源描述
        example: GitHub Actions 构建通知
        type: string
      id:
        description: 来源 ID
        example: 1
        type: integer
      last_used_at:
        description: 最后使用时间（Unix 毫秒时间戳，0 表示未使用）
        example: 1714636800000
        type: integer
      name:
        description: 来源名称
        example: CI Pipeline
        type: string
      status:
        description: 状态：1-启用 2-停用 -1-已删除
        example: 1
        type: integer
      token:
        description: 推送 Token
        example: src0196a3b2c4d50000a1b2c3d4e5f67890
        type: string
      updated_at:
        description: 更新时间（Unix 毫秒时间戳）
        example: 1714636800000
        type: integer
      user_id:
        description: 所属用户 ID
        example: 1
        type: integer
    type: object
  octotify_internal_handler_dto.SourceDetailResponse:
    description: 来源详情及其已绑定的有效渠道列表
    properties:
      channels:
        description: 已绑定渠道列表
        items:
          $ref: '#/definitions/octotify_internal_handler_dto.ChannelDTO'
        type: array
      source:
        allOf:
        - $ref: '#/definitions/octotify_internal_handler_dto.SourceDetailDTO'
        description: 来源详情
    type: object
  octotify_internal_handler_dto.SourceTokenResponse:
    description: 查看或重置来源令牌时返回的 Token 值
    properties:
      token:
        description: 推送 Token
        example: src0196a3b2c4d50000a1b2c3d4e5f67890
        type: string
    type: object
...
  octotify_internal_handler_dto.UpdateSourceReq:
    description: 编辑已有消息来源的名称和描述
    properties:
      description:
        description: 来源描述，最多 512 个字符
        example: GitHub Actions 构建通知
        maxLength: 512
        type: string
      name:
        description: 来源名称，1-128 个字符
        example: CI Pipeline
        maxLength: 128
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  octotify_internal_handler_dto.VerifyPasswordReq:
    description: 需要密码二次验证的接口（查看令牌、重置令牌、停用/启用/删除来源）
    properties:
      password:
        description: 用户密码
        example: P@ssw0rd123
        type: string
    required:
    - password
    type: object
  octotify_pkg_response.PageResult:
    description: 所有分页接口返回的数据结构
    properties:
      list:
        description: 数据列表
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页条数
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
  octotify_pkg_response.Response:
    description: 所有 API 接口返回的统一格式，包含业务状态码、消息和数据
    properties:
      code:
        description: 业务状态码，0 表示成功，非 0 表示失败
        type: integer
      data:
        description: 业务数据，成功时返回业务数据，失败时可能为空或错误详情
      msg:
        description: 提示信息，成功时为"请求成功"，失败时为错误描述
        type: string
    type: object
host: localhost:34123
info:
  contact:
    name: OctoTify API Support
    url: https://github.com/OctoTify
  description: |-
    OctoTify 是一个消息总线平台，支持多种消息来源和推送渠道。
    核心功能：消息来源管理、推送渠道管理、消息推送与记录。
    本文档为前端开发人员提供完整的 API 接口说明，包含请求参数、响应格式、错误码等详细信息。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: OctoTify API
  version: "1.0"
paths:
...
  /sources:
    get:
      consumes:
      - application/json
      description: |-
        分页查询当前用户的所有消息来源列表，返回来源基本信息（不包含 Token）。
        ## 使用场景
        1. 来源管理页面展示来源列表
        2. 下拉选择器获取来源选项
        ## 注意事项
        - 列表按创建时间倒序排列
        - 返回的来源不包含 Token 字段
        - 已删除的来源不会出现在列表中
        ## 错误码说明
        - 100001: 未提供认证令牌
      parameters:
      - default: 1
        description: 页码，从 1 开始，默认 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: 每页条数，默认 20，最大 100
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/octotify_pkg_response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/octotify_pkg_response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/octotify_internal_handler_dto.SourceDTO'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 查看消息来源列表
      tags:
      - 消息来源管理
    post:
      consumes:
      - application/json
      description: |-
        创建一个新的消息来源，系统自动生成 Source Token，用于外部系统推送消息。
        ## 使用场景
        1. CI/CD 流水线创建为消息来源
        2. 监控系统创建为消息来源
        3. 业务系统创建为消息来源
        ## 注意事项
        - Source Token 仅在创建时返回一次，请妥善保存
        - 如忘记 Token，可通过查看令牌接口获取（需密码验证）
        - 来源创建后默认启用状态
        - 每个来源可绑定多个推送渠道
        ## 错误码说明
        - 100001: 未提供认证令牌
        - 120100: 来源名称已存在
        - 120101: 创建来源失败
      parameters:
      - description: 创建消息来源请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/octotify_internal_handler_dto.CreateSourceReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/octotify_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/octotify_internal_handler_dto.SourceDTO'
              type: object
      security:
      - BearerAuth: []
      summary: 创建消息来源
      tags:
      - 消息来源管理
  /sources/{id}:
    delete:
      consumes:
      - application/json
      description: |-
        软删除指定消息来源及其关联渠道关系，需要密码二次验证。删除后来源和所有关联数据不可恢复。
        ## 使用场景
        1. 永久删除不再使用的来源
        2. 清理测试数据
        ## 注意事项
        - 删除是软删除，数据库中标记为已删除状态
        - 删除后该来源的 Token 立即失效
        - 关联的渠道关系也会被删除
        - 只能删除自己创建的来源
        - 注意：DELETE 请求需在 Body 中传递密码（{"password": "xxx"}），主流网关均支持此用法
        ## 错误码说明
        - 100001: 未提供认证令牌
        - 110200: 密码错误
        - 120102: 来源不存在
        - 120103: 无权限操作
        - 120107: 删除来源失败
      parameters:
      - description: 来源ID，在路径中传递
        format: int64
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 密码验证请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/octotify_internal_handler_dto.VerifyPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/octotify_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除消息来源
      tags:
      - 消息来源管理
    get:
      consumes:
      - application/json
      description: |-
        查询指定消息来源的详细信息，包含来源 Token、使用时间和已绑定的有效渠道列表。
        ## 使用场景
        1. 来源详情页展示
        2. 查看来源绑定的渠道
        3. 查看来源的推送 Token
        ## 注意事项
        - 返回的 Token 可用于推送消息接口
        - 渠道列表只包含启用状态的渠道
        - 只能查看自己创建的来源
        ## 错误码说明
        - 100001: 未提供认证令牌
        - 120102: 来源不存在
        - 120103: 无权限查看
      parameters:
      - description: 来源ID，在路径中传递
        format: int64
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/octotify_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/octotify_internal_handler_dto.SourceDetailResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 查看来源详情
      tags:
      - 消息来源管理
    put:
      consumes:
      - application/json
      description: |-
        编辑消息来源的名称和描述，不影响已绑定的渠道和推送功能。
        ## 使用场景
        1. 修改来源名称使其更易识别
        2. 更新来源描述信息
        ## 注意事项
        - 只能编辑自己创建的来源
        - 名称不能与其他来源重复
        - 编辑后不影响 Source Token
        ## 错误码说明
        - 100001: 未提供认证令牌
        - 120100: 来源名称已存在
        - 120102: 来源不存在
        - 120103: 无权限编辑
      parameters:
      - description: 来源ID，在路径中传递
        format: int64
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 编辑消息来源请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/octotify_internal_handler_dto.UpdateSourceReq'
      produces:
      - application/json
      responses:
        "200":
          description: 编辑成功
          schema:
            $ref: '#/definitions/octotify_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 编辑消息来源
      tags:
      - 消息来源管理
  /sources/{id}/disable:
    put:
      consumes:
      - application/json
      description: |-
        停用指定消息来源，需要密码二次验证，停用后该来源无法推送消息。
        ## 使用场景
        1. 临时停止某个来源的消息推送
        2. 来源不再使用时停用而非删除
        ## 注意事项
        - 停用后使用该 Token 推送的消息会被拒绝
        - 已推送的历史消息不受影响
        - 可以随时重新启用
        - 只能停用自己创建的来源
        ## 错误码说明
        - 100001: 未提供认证令牌
        - 110200: 密码错误
        - 120102: 来源不存在
        - 120103: 无权限操作
        - 120105: 来源已停用
      parameters:
      - description: 来源ID，在路径中传递
        format: int64
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 密码验证请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/octotify_internal_handler_dto.VerifyPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 停用成功
          schema:
            $ref: '#/definitions/octotify_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 停用消息来源
      tags:
      - 消息来源管理
  /sources/{id}/enable:
    put:
      consumes:
      - application/json
      description: |-
        启用指定消息来源，需要密码二次验证，恢复消息推送功能。
        ## 使用场景
        1. 重新启用之前停用的来源
        2. 恢复某个来源的消息推送能力
        ## 注意事项
        - 只能启用已停用的来源
        - 启用后该来源的 Token 立即生效
        - 只能启用自己创建的来源
        ## 错误码说明
        - 100001: 未提供认证令牌
        - 110200: 密码错误
        - 120102: 来源不存在
        - 120103: 无权限操作
        - 120106: 来源已启用
      parameters:
      - description: 来源ID，在路径中传递
        format: int64
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 密码验证请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/octotify_internal_handler_dto.VerifyPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 启用成功
          schema:
            $ref: '#/definitions/octotify_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 启用消息来源
      tags:
      - 消息来源管理
  /sources/{id}/token:
    post:
      consumes:
      - application/json
      description: |-
        查询指定消息来源的 Token，需要密码二次验证以确保安全性。
        ## 使用场景
        1. 忘记 Token 时查看
        2. 配置外部系统推送时获取 Token
        ## 注意事项
        - 必须提供正确的用户密码才能查看
        - Token 格式为 src{uuid}，例如：src0196a3b2c4d50000a1b2c3d4e5f67890
        - 只能查看自己创建的来源
        ## 错误码说明
        - 100001: 未提供认证令牌
        - 110200: 密码错误
        - 120102: 来源不存在
        - 120103: 无权限查看
      parameters:
      - description: 来源ID，在路径中传递
        format: int64
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 密码验证请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/octotify_internal_handler_dto.VerifyPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 查看成功
          schema:
            allOf:
            - $ref: '#/definitions/octotify_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/octotify_internal_handler_dto.SourceTokenResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 查看来源令牌
      tags:
      - 消息来源管理
    put:
      consumes:
      - application/json
      description: |-
        重置指定消息来源的 Token，需要密码二次验证，旧 Token 立即失效。
        ## 使用场景
        1. Token 泄露时紧急重置
        2. 定期更换 Token 以提高安全性
        ## 注意事项
        - 重置后旧 Token 立即失效，使用旧 Token 推送会失败
        - 必须更新所有使用旧 Token 的外部系统配置
        - 新 Token 仅在重置时返回一次，请妥善保存
        - 只能重置自己创建的来源
        ## 错误码说明
        - 100001: 未提供认证令牌
        - 110200: 密码错误
        - 120102: 来源不存在
        - 120103: 无权限操作
        - 120104: 重置令牌失败
      parameters:
      - description: 来源ID，在路径中传递
        format: int64
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 密码验证请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/octotify_internal_handler_dto.VerifyPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            allOf:
            - $ref: '#/definitions/octotify_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/octotify_internal_handler_dto.SourceTokenResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 重置来源令牌
      tags:
      - 消息来源管理
...
securityDefinitions:
  BearerAuth:
    description: 输入格式：Bearer {access_token}，例如：Bearer eyJhbGciOiJSUzI1NiIs...
    in: header
    name: Authorization
    type: apiKey
  SourceTokenAuth:
    description: 推送消息时使用 Source Token，输入格式：Bearer src{uuid}，例如：Bearer src0196a3b2c4d50000a1b2c3d4e5f67890
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...